fix opt-lambdas with rest arguments
An opt-lambda is expanded to a case lambda that calls a generated lambda, whose
last positional parameter corresponds to the original rest
parameter. Previously, the type for the former was the same as the type user
specified for the latter, which caused unsoundness problems. The fix changed the
type of the last positional parameter from `ty` to `(Listof ty)`and also ensured
the type label on identifier for the original rest paramter that used at the LHS
is ignored